Correct SW1 positions and PWR connections
SW1 was well placed and wired but the text in copper layer that explains
the two posible positions was wrong.

PWR connector was wrong wired. Now all the circuit on the board is
supplied by the PUSHBUTTON terminal. In this way, when either the
Arduino, either the fuel gauge shut down the UGV, all the circuitry will
shut down.