Prepare Silk Layer in order to be able to manufacture
This 'manu2' branch was created in order to have the 3 projects
on the same folder.

The silk layer was created and adapted to give it to an external
manufacturer.
Nearly the same information was created in the copper layer in order to
have this information also in the case the PCB's were manufactured in
the DTE.
The Mask layer was also adapted to give it to the external manufacturer.